packing partner recipe, but not binding
